migrate existing tests
